phase-4-kubernetes-deployment/backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI backend for Evolution of Todo Phase 2"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pyjwt[crypto]>=2.9.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.0.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.9",
    "pillow>=11.0.0",
    "openai>=1.0.0",
    "openai-agents>=0.0.3",
    "openai-chatkit>=0.1.0",
    "sse-starlette>=2.0.0",
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
